send ftm

+++ Quiiiz.Peon/Configuration/Blockchain.cs
@@ -7,5 +7,5 @@ public record class Blockchain
     public required string Seed { get; init; }
     public required string Password { get; init; }
     public required Uri Node { get; init; }
-    public required BigInteger ChainId { get; init; }
+    public required long ChainId { get; init; }
 }
